+++ HISTORY.md
@@ -14,7 +14,7 @@
 - `OAuth2Client.authorization_code()` now accepts an `AuthorizationResponse` as parameter, and will
   use it to include all necessary parameters for the Authorization Code Grant.
 - removed `OAuth2Client.authorization_code_pkce()`
-- Added `ClientSecretJwt` and `PrivateKeyJwt` aliases to `ClientSecretJWT` and `PrivateKeyJWT`, for consistency with `jwskate`.
+- Renamed `ClientSecretJWT` and `PrivateKeyJWT` to `ClientSecretJwt` and `PrivateKeyJwt`, for consistency with `jwskate`.
 - Methods from `requests_oauth2client.utils` are no longer exposed in top-level module.
 - Renamed base class `ClientAuthenticationMethod` to `BaseClientAuthenticationMethod`.
 - Introduced a default timeout in `ApiClient`